refactor(mcp): extract transport client internals
- move stdio client lifecycle and request dispatch into module
- move streamable HTTP SSE parsing/listener helpers to transport
- route operations through shared HTTP context and request hooks
- persist negotiated protocol version separately in server state
- update README MCP module descriptions for transport responsibilities
