+++ src/mcp/client/mod.rs
@@ -9,28 +9,24 @@ pub use operations::{
     execute_prompt, execute_resource_list, execute_resource_read, execute_resource_template_list,
     execute_tool_call, send_client_error, send_client_result,
 };
-use rust_mcp_schema::schema_utils::{
-    ClientMessage, FromMessage, MessageFromClient, NotificationFromClient, RequestFromClient,
-    ResultFromClient, ServerMessage,
-};
+use rust_mcp_schema::schema_utils::{RequestFromClient, ServerMessage};
 use rust_mcp_schema::{
     ClientCapabilities, ClientSampling, Implementation, InitializeRequestParams, InitializeResult,
     ListPromptsResult, ListResourceTemplatesResult, ListResourcesResult, ListToolsResult,
-    PaginatedRequestParams, RequestId, RpcError, ServerCapabilities,
+    PaginatedRequestParams, RpcError, ServerCapabilities,
 };
 use std::collections::HashMap;
-use std::sync::atomic::{AtomicI64, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
-use tokio::io::{AsyncBufReadExt, AsyncWriteExt, BufReader};
-use tokio::process::{ChildStdin, Command};
-use tokio::sync::{mpsc, oneshot, Mutex, Notify, RwLock};
-use tracing::debug;
+use tokio::sync::mpsc;
 
 mod operations;
 mod protocol;
 mod transport_http;
 mod transport_stdio;
+
+use transport_http::StreamableHttpContext;
+use transport_stdio::StdioClient;
 
 const MCP_MAX_TOOL_LIST: usize = 100;
 const MCP_STARTUP_CONCURRENCY_LIMIT: usize = 3;
@@ -97,513 +93,6 @@ fn apply_streamable_http_protocol_version_header(
 const STDIO_REQUEST_TIMEOUT_SECONDS: u64 = 60;
 const STDIO_SAMPLING_TIMEOUT_MULTIPLIER: u64 = 5;
 
-struct StdioClient {
-    stdin: Mutex<ChildStdin>,
-    pending: Arc<Mutex<HashMap<RequestId, oneshot::Sender<ServerMessage>>>>,
-    next_request_id: AtomicI64,
-    server_details: RwLock<Option<rust_mcp_schema::InitializeResult>>,
-    server_id: String,
-    request_tx: Option<mpsc::UnboundedSender<McpServerRequest>>,
-    activity_notify: Arc<Notify>,
-    inflight_server_requests: Arc<AtomicI64>,
-}
-
-#[derive(Default)]
-struct SseLineBuffer {
-    buffer: Vec<u8>,
-}
-
-impl SseLineBuffer {
-    fn push(&mut self, chunk: &[u8]) -> Vec<String> {
-        self.buffer.extend_from_slice(chunk);
-        self.drain_lines(false)
-    }
-
-    fn finish(&mut self) -> Vec<String> {
-        self.drain_lines(true)
-    }
-
-    fn drain_lines(&mut self, flush: bool) -> Vec<String> {
-        let mut lines = Vec::new();
-        let mut search_index = 0;
-
-        while let Some(relative_pos) = self.buffer[search_index..].iter().position(|b| *b == b'\n')
-        {
-            let newline_index = search_index + relative_pos;
-            let mut line_end = newline_index;
-            if line_end > search_index && self.buffer[line_end - 1] == b'\r' {
-                line_end -= 1;
-            }
-
-            let line_bytes = &self.buffer[search_index..line_end];
-            if let Ok(text) = std::str::from_utf8(line_bytes) {
-                let trimmed = text.trim();
-                if !trimmed.is_empty() {
-                    lines.push(trimmed.to_string());
-                }
-            }
-
-            search_index = newline_index + 1;
-        }
-
-        if flush {
-            if let Ok(text) = std::str::from_utf8(&self.buffer[search_index..]) {
-                let trimmed = text.trim();
-                if !trimmed.is_empty() {
-                    lines.push(trimmed.to_string());
-                }
-            }
-            self.buffer.clear();
-        } else if search_index > 0 {
-            self.buffer.drain(..search_index);
-        }
-
-        lines
-    }
-}
-
-impl StdioClient {
-    async fn connect(
-        server_id: String,
-        config: &McpServerConfig,
-        request_tx: Option<mpsc::UnboundedSender<McpServerRequest>>,
-    ) -> Result<Arc<Self>, String> {
-        let command = require_stdio_command(config)?;
-        let args = stdio_args(config);
-        debug!(command = %command, args = ?args, "Starting MCP stdio server");
-        let mut cmd = Command::new(command);
-        cmd.args(args)
-            .stdin(std::process::Stdio::piped())
-            .stdout(std::process::Stdio::piped())
-            .stderr(std::process::Stdio::piped());
-
-        if let Some(env) = stdio_env(config) {
-            cmd.envs(env);
-        }
-
-        let mut child = cmd.spawn().map_err(|err| err.to_string())?;
-        let stdin = child
-            .stdin
-            .take()
-            .ok_or_else(|| "Unable to retrieve stdin.".to_string())?;
-        let stdout = child
-            .stdout
-            .take()
-            .ok_or_else(|| "Unable to retrieve stdout.".to_string())?;
-        let stderr = child
-            .stderr
-            .take()
-            .ok_or_else(|| "Unable to retrieve stderr.".to_string())?;
-
-        let pending: Arc<Mutex<HashMap<RequestId, oneshot::Sender<ServerMessage>>>> =
-            Arc::new(Mutex::new(HashMap::new()));
-        let client = Arc::new(Self {
-            stdin: Mutex::new(stdin),
-            pending: pending.clone(),
-            next_request_id: AtomicI64::new(0),
-            server_details: RwLock::new(None),
-            server_id,
-            request_tx,
-            activity_notify: Arc::new(Notify::new()),
-            inflight_server_requests: Arc::new(AtomicI64::new(0)),
-        });
-
-        Self::spawn_stdout_reader(
-            pending.clone(),
-            stdout,
-            client.server_id.clone(),
-            client.request_tx.clone(),
-            client.activity_notify.clone(),
-            client.inflight_server_requests.clone(),
-        );
-        Self::spawn_stderr_drain(stderr);
-
-        tokio::spawn(async move {
-            let _ = child.wait().await;
-            let mut pending = pending.lock().await;
-            pending.clear();
-        });
-
-        Ok(client)
-    }
-
-    fn spawn_stdout_reader(
-        pending: Arc<Mutex<HashMap<RequestId, oneshot::Sender<ServerMessage>>>>,
-        stdout: tokio::process::ChildStdout,
-        server_id: String,
-        request_tx: Option<mpsc::UnboundedSender<McpServerRequest>>,
-        activity_notify: Arc<Notify>,
-        inflight_server_requests: Arc<AtomicI64>,
-    ) {
-        tokio::spawn(async move {
-            let mut reader = BufReader::new(stdout).lines();
-            while let Ok(Some(line)) = reader.next_line().await {
-                let value = match serde_json::from_str::<serde_json::Value>(&line) {
-                    Ok(value) => value,
-                    Err(_) => continue,
-                };
-                if let Some(items) = value.as_array() {
-                    for item in items {
-                        if let Ok(message) = serde_json::from_value::<ServerMessage>(item.clone()) {
-                            Self::dispatch_message(
-                                &pending,
-                                message,
-                                &server_id,
-                                request_tx.as_ref(),
-                                &activity_notify,
-                                &inflight_server_requests,
-                            )
-                            .await;
-                        }
-                    }
-                } else if let Ok(message) = serde_json::from_value::<ServerMessage>(value) {
-                    Self::dispatch_message(
-                        &pending,
-                        message,
-                        &server_id,
-                        request_tx.as_ref(),
-                        &activity_notify,
-                        &inflight_server_requests,
-                    )
-                    .await;
-                }
-            }
-        });
-    }
-
-    fn spawn_stderr_drain(stderr: tokio::process::ChildStderr) {
-        tokio::spawn(async move {
-            let mut reader = BufReader::new(stderr).lines();
-            while let Ok(Some(_)) = reader.next_line().await {}
-        });
-    }
-
-    async fn dispatch_message(
-        pending: &Arc<Mutex<HashMap<RequestId, oneshot::Sender<ServerMessage>>>>,
-        message: ServerMessage,
-        server_id: &str,
-        request_tx: Option<&mpsc::UnboundedSender<McpServerRequest>>,
-        activity_notify: &Notify,
-        inflight_server_requests: &AtomicI64,
-    ) {
-        match &message {
-            ServerMessage::Response(response) => {
-                debug!(
-                    server_id = %server_id,
-                    response_id = ?response.id,
-                    "Received MCP stdio response"
-                );
-                if let Some(tx) = pending.lock().await.remove(&response.id) {
-                    let _ = tx.send(message);
-                }
-            }
-            ServerMessage::Error(error) => {
-                debug!(
-                    server_id = %server_id,
-                    error_id = ?error.id,
-                    error_code = error.error.code,
-                    "Received MCP stdio error"
-                );
-                if let Some(id) = error.id.as_ref() {
-                    if let Some(tx) = pending.lock().await.remove(id) {
-                        let _ = tx.send(message);
-                    }
-                }
-            }
-            ServerMessage::Request(request) => {
-                let inflight = inflight_server_requests.fetch_add(1, Ordering::SeqCst) + 1;
-                debug!(
-                    server_id = %server_id,
-                    method = %request.method(),
-                    request_id = ?request.request_id(),
-                    inflight_server_requests = inflight,
-                    "Received MCP stdio request"
-                );
-                activity_notify.notify_waiters();
-                if let Some(tx) = request_tx {
-                    let _ = tx.send(McpServerRequest {
-                        server_id: server_id.to_string(),
-                        request: request.clone(),
-                    });
-                }
-            }
-            ServerMessage::Notification(_) => {
-                debug!(server_id = %server_id, "Received MCP stdio notification");
-                activity_notify.notify_waiters();
-            }
-        }
-    }
-}
-
-impl StdioClient {
-    fn timeout_for_wait(&self) -> tokio::time::Duration {
-        let inflight = self.inflight_server_requests.load(Ordering::SeqCst);
-        let multiplier = if inflight > 0 {
-            STDIO_SAMPLING_TIMEOUT_MULTIPLIER
-        } else {
-            1
-        };
-        tokio::time::Duration::from_secs(STDIO_REQUEST_TIMEOUT_SECONDS * multiplier)
-    }
-
-    fn decrement_inflight(&self) -> i64 {
-        let mut current = self.inflight_server_requests.load(Ordering::SeqCst);
-        while current > 0 {
-            match self.inflight_server_requests.compare_exchange(
-                current,
-                current - 1,
-                Ordering::SeqCst,
-                Ordering::SeqCst,
-            ) {
-                Ok(_) => return current - 1,
-                Err(next) => current = next,
-            }
-        }
-        current
-    }
-}
-
-impl StdioClient {
-    fn next_request_id(&self) -> RequestId {
-        let id = self.next_request_id.fetch_add(1, Ordering::SeqCst);
-        RequestId::Integer(id)
-    }
-
-    async fn send_request(&self, request: RequestFromClient) -> Result<ServerMessage, String> {
-        let request_id = self.next_request_id();
-        debug!(request_id = ?request_id, "Sending MCP stdio request");
-        let message = ClientMessage::from_message(
-            MessageFromClient::RequestFromClient(request),
-            Some(request_id.clone()),
-        )
-        .map_err(|err| err.to_string())?;
-
-        let (tx, rx) = oneshot::channel();
-        {
-            let mut pending = self.pending.lock().await;
-            pending.insert(request_id.clone(), tx);
-        }
-
-        let payload = serde_json::to_string(&message).map_err(|err| err.to_string())?;
-        let lock_timeout = tokio::time::Duration::from_secs(10);
-        let write_timeout = tokio::time::Duration::from_secs(10);
-        let lock_start = std::time::Instant::now();
-        debug!(request_id = ?request_id, "Awaiting MCP stdio stdin lock for request");
-        let mut stdin = match tokio::time::timeout(lock_timeout, self.stdin.lock()).await {
-            Ok(stdin) => stdin,
-            Err(_) => {
-                return Err("Timed out waiting for MCP stdio stdin lock.".to_string());
-            }
-        };
-        debug!(
-            request_id = ?request_id,
-            elapsed_ms = lock_start.elapsed().as_millis(),
-            bytes = payload.len(),
-            "Writing MCP stdio request"
-        );
-        tokio::time::timeout(write_timeout, stdin.write_all(payload.as_bytes()))
-            .await
-            .map_err(|_| "Timed out writing MCP stdio request.".to_string())?
-            .map_err(|err| err.to_string())?;
-        tokio::time::timeout(write_timeout, stdin.write_all(b"\n"))
-            .await
-            .map_err(|_| "Timed out writing MCP stdio request newline.".to_string())?
-            .map_err(|err| err.to_string())?;
-        tokio::time::timeout(write_timeout, stdin.flush())
-            .await
-            .map_err(|_| "Timed out flushing MCP stdio request.".to_string())?
-            .map_err(|err| err.to_string())?;
-        debug!(request_id = ?request_id, "MCP stdio request sent");
-        drop(stdin);
-
-        let mut timeout = self.timeout_for_wait();
-        let mut deadline = tokio::time::Instant::now() + timeout;
-        let mut rx = rx;
-        loop {
-            let now = tokio::time::Instant::now();
-            if now >= deadline {
-                debug!(
-                    request_id = ?request_id,
-                    timeout_secs = timeout.as_secs(),
-                    "MCP stdio request timed out (deadline reached)"
-                );
-                return Err("MCP stdio request timed out.".to_string());
-            }
-            let remaining = deadline - now;
-            debug!(
-                request_id = ?request_id,
-                remaining_ms = remaining.as_millis(),
-                inflight_server_requests = self.inflight_server_requests.load(Ordering::SeqCst),
-                timeout_secs = timeout.as_secs(),
-                "Awaiting MCP stdio response"
-            );
-            tokio::select! {
-                result = &mut rx => {
-                    return match result {
-                        Ok(message) => {
-                            debug!(request_id = ?request_id, "MCP stdio response received");
-                            Ok(message)
-                        }
-                        Err(_) => {
-                            debug!(request_id = ?request_id, "MCP stdio response channel closed");
-                            Err("MCP stdio response channel closed.".to_string())
-                        }
-                    };
-                }
-                _ = tokio::time::sleep(remaining) => {
-                    debug!(
-                        request_id = ?request_id,
-                        timeout_secs = timeout.as_secs(),
-                        "MCP stdio request timed out (sleep elapsed)"
-                    );
-                    return Err("MCP stdio request timed out.".to_string());
-                }
-                _ = self.activity_notify.notified() => {
-                    timeout = self.timeout_for_wait();
-                    debug!(
-                        request_id = ?request_id,
-                        timeout_secs = timeout.as_secs(),
-                        inflight_server_requests = self.inflight_server_requests.load(Ordering::SeqCst),
-                        "MCP stdio timeout reset after server activity"
-                    );
-                    deadline = tokio::time::Instant::now() + timeout;
-                }
-            }
-        }
-    }
-
-    async fn send_notification(&self, notification: NotificationFromClient) -> Result<(), String> {
-        let message = ClientMessage::from_message(
-            MessageFromClient::NotificationFromClient(notification),
-            None,
-        )
-        .map_err(|err| err.to_string())?;
-        let payload = serde_json::to_string(&message).map_err(|err| err.to_string())?;
-        let lock_timeout = tokio::time::Duration::from_secs(10);
-        let write_timeout = tokio::time::Duration::from_secs(10);
-        let lock_start = std::time::Instant::now();
-        debug!(server_id = %self.server_id, "Awaiting MCP stdio stdin lock for notification");
-        let mut stdin = match tokio::time::timeout(lock_timeout, self.stdin.lock()).await {
-            Ok(stdin) => stdin,
-            Err(_) => {
-                return Err("Timed out waiting for MCP stdio stdin lock.".to_string());
-            }
-        };
-        debug!(
-            server_id = %self.server_id,
-            elapsed_ms = lock_start.elapsed().as_millis(),
-            bytes = payload.len(),
-            "Writing MCP stdio notification"
-        );
-        tokio::time::timeout(write_timeout, stdin.write_all(payload.as_bytes()))
-            .await
-            .map_err(|_| "Timed out writing MCP stdio notification.".to_string())?
-            .map_err(|err| err.to_string())?;
-        tokio::time::timeout(write_timeout, stdin.write_all(b"\n"))
-            .await
-            .map_err(|_| "Timed out writing MCP stdio notification newline.".to_string())?
-            .map_err(|err| err.to_string())?;
-        tokio::time::timeout(write_timeout, stdin.flush())
-            .await
-            .map_err(|_| "Timed out flushing MCP stdio notification.".to_string())?
-            .map_err(|err| err.to_string())?;
-        debug!(server_id = %self.server_id, "MCP stdio notification sent");
-        Ok(())
-    }
-
-    async fn initialize(
-        &self,
-        details: InitializeRequestParams,
-    ) -> Result<InitializeResult, String> {
-        let response = self
-            .send_request(RequestFromClient::InitializeRequest(details))
-            .await?;
-        let result = protocol::parse_initialize_result(response)?;
-        *self.server_details.write().await = Some(result.clone());
-        self.send_notification(NotificationFromClient::InitializedNotification(None))
-            .await?;
-        Ok(result)
-    }
-
-    async fn send_result(
-        &self,
-        request_id: RequestId,
-        result: ResultFromClient,
-    ) -> Result<(), String> {
-        debug!(
-            server_id = %self.server_id,
-            request_id = ?request_id,
-            "Preparing MCP stdio result"
-        );
-        let message = ClientMessage::from_message(
-            MessageFromClient::ResultFromClient(result),
-            Some(request_id.clone()),
-        )
-        .map_err(|err| err.to_string())?;
-        let result = self.send_client_message(&message).await;
-        if result.is_ok() {
-            let inflight = self.decrement_inflight();
-            debug!(
-                request_id = ?request_id,
-                inflight_server_requests = inflight,
-                "Sent MCP stdio result"
-            );
-            self.activity_notify.notify_waiters();
-        }
-        result
-    }
-
-    async fn send_error(&self, request_id: RequestId, error: RpcError) -> Result<(), String> {
-        debug!(
-            server_id = %self.server_id,
-            request_id = ?request_id,
-            "Preparing MCP stdio error response"
-        );
-        let message =
-            ClientMessage::from_message(MessageFromClient::Error(error), Some(request_id.clone()))
-                .map_err(|err| err.to_string())?;
-        let result = self.send_client_message(&message).await;
-        if result.is_ok() {
-            let inflight = self.decrement_inflight();
-            debug!(
-                request_id = ?request_id,
-                inflight_server_requests = inflight,
-                "Sent MCP stdio error response"
-            );
-            self.activity_notify.notify_waiters();
-        }
-        result
-    }
-
-    async fn send_client_message(&self, message: &ClientMessage) -> Result<(), String> {
-        let lock_timeout = tokio::time::Duration::from_secs(10);
-        debug!(server_id = %self.server_id, "Awaiting MCP stdio stdin lock");
-        let payload = serde_json::to_string(message).map_err(|err| err.to_string())?;
-        let mut stdin = match tokio::time::timeout(lock_timeout, self.stdin.lock()).await {
-            Ok(stdin) => stdin,
-            Err(_) => {
-                return Err("Timed out waiting for MCP stdio stdin lock.".to_string());
-            }
-        };
-        debug!(server_id = %self.server_id, bytes = payload.len(), "Writing MCP stdio client message");
-        let write_timeout = tokio::time::Duration::from_secs(10);
-        tokio::time::timeout(write_timeout, stdin.write_all(payload.as_bytes()))
-            .await
-            .map_err(|_| "Timed out writing MCP stdio client message.".to_string())?
-            .map_err(|err| err.to_string())?;
-        tokio::time::timeout(write_timeout, stdin.write_all(b"\n"))
-            .await
-            .map_err(|_| "Timed out writing MCP stdio newline.".to_string())?
-            .map_err(|err| err.to_string())?;
-        tokio::time::timeout(write_timeout, stdin.flush())
-            .await
-            .map_err(|_| "Timed out flushing MCP stdio client message.".to_string())?
-            .map_err(|err| err.to_string())?;
-        debug!(server_id = %self.server_id, "MCP stdio client message sent");
-        Ok(())
-    }
-}
-
 #[derive(Clone)]
 pub struct McpServerState {
     pub config: McpServerConfig,
@@ -616,6 +105,7 @@ pub struct McpServerState {
     pub session_id: Option<String>,
     pub auth_header: Option<String>,
     pub server_details: Option<InitializeResult>,
+    pub negotiated_protocol_version: Option<String>,
     pub streamable_http_request_id: u64,
     pub event_listener_started: bool,
     http_client: Option<reqwest::Client>,
@@ -635,6 +125,7 @@ impl McpServerState {
             session_id: None,
             auth_header: None,
             server_details: None,
+            negotiated_protocol_version: None,
             streamable_http_request_id: 0,
             event_listener_started: false,
             http_client: None,
@@ -680,6 +171,7 @@ impl McpServerState {
         self.session_id = None;
         self.auth_header = None;
         self.server_details = None;
+        self.negotiated_protocol_version = None;
         self.streamable_http_request_id = 0;
         self.event_listener_started = false;
         self.http_client = None;
@@ -747,19 +239,16 @@ impl McpClientManager {
         self.server_request_tx = Some(sender);
     }
 
-    fn ensure_http_client(&mut self, id: &str) -> Result<reqwest::Client, String> {
+    fn ensure_http_client(&mut self, id: &str) -> Result<(), String> {
         let Some(server) = self.server_mut(id) else {
             return Err("Unknown MCP server".to_string());
         };
-
-        if let Some(client) = server.http_client.clone() {
-            return Ok(client);
+        if server.http_client.is_none() {
+            let client = build_mcp_http_client()
+                .map_err(|err| format!("Failed to build HTTP client: {err}"))?;
+            server.http_client = Some(client);
         }
-
-        let client =
-            build_mcp_http_client().map_err(|err| format!("Failed to build HTTP client: {err}"))?;
-        server.http_client = Some(client.clone());
-        Ok(client)
+        Ok(())
     }
 
     fn apply_list_fetch<T>(
@@ -1049,7 +538,7 @@ impl McpClientManager {
                             if let Some(server) = self.server_mut(id) {
                                 if !server.event_listener_started {
                                     server.event_listener_started = true;
-                                    spawn_streamable_http_listener(
+                                    transport_http::spawn_streamable_http_listener(
                                         server
                                             .http_client
                                             .clone()
@@ -1061,10 +550,13 @@ impl McpClientManager {
                                         tx,
                                         Some(protocol::effective_protocol_version(
                                             &server.config,
-                                            server
-                                                .server_details
-                                                .as_ref()
-                                                .map(|details| details.protocol_version.as_str()),
+                                            server.negotiated_protocol_version.as_deref().or_else(
+                                                || {
+                                                    server.server_details.as_ref().map(|details| {
+                                                        details.protocol_version.as_str()
+                                                    })
+                                                },
+                                            ),
                                         )),
                                     );
                                 }
@@ -1108,6 +600,8 @@ impl McpClientManager {
                 match client.initialize(client_details).await {
                     Ok(details) => {
                         if let Some(server) = self.server_mut(id) {
+                            server.negotiated_protocol_version =
+                                Some(details.protocol_version.clone());
                             server.server_details = Some(details);
                         }
                     }
@@ -1343,115 +837,17 @@ impl McpClientManager {
     }
 
     async fn ensure_streamable_http_session(&mut self, id: &str) -> Result<(), String> {
-        let needs_init = self
-            .server(id)
-            .map(|server| server.session_id.is_none())
-            .unwrap_or(false);
-        if !needs_init {
-            return Ok(());
-        }
+        self.ensure_http_client(id)?;
 
-        let client_details = self
-            .server(id)
-            .map(|server| client_details_for(&server.config))
-            .ok_or_else(|| "Unknown MCP server".to_string())?;
-
-        let response = self
-            .send_streamable_http_request(id, RequestFromClient::InitializeRequest(client_details))
-            .await?;
-        let initialize = protocol::parse_initialize_result(response)?;
-
-        if let Some(server) = self.server_mut(id) {
-            server.server_details = Some(initialize);
-        }
-
-        if self
-            .server(id)
-            .and_then(|server| server.session_id.as_deref())
-            .is_none()
-        {
-            return Err("Missing session id on initialize response.".to_string());
-        }
-
-        self.send_streamable_http_notification(
-            id,
-            NotificationFromClient::InitializedNotification(None),
-        )
-        .await?;
-
-        Ok(())
-    }
-
-    async fn send_streamable_http_notification(
-        &mut self,
-        id: &str,
-        notification: NotificationFromClient,
-    ) -> Result<(), String> {
-        let message = ClientMessage::from_message(
-            MessageFromClient::NotificationFromClient(notification),
-            None,
-        )
-        .map_err(|err| err.to_string())?;
-
-        self.send_streamable_http_client_message(id, message).await
-    }
-
-    async fn send_streamable_http_client_message(
-        &mut self,
-        id: &str,
-        message: ClientMessage,
-    ) -> Result<(), String> {
-        let payload = serde_json::to_string(&message).map_err(|err| err.to_string())?;
-
-        let (base_url, auth_header, session_id, protocol_version) = {
-            let Some(server) = self.server(id) else {
-                return Err("Unknown MCP server".to_string());
-            };
-            (
-                require_http_base_url(&server.config)?,
-                server.auth_header.clone(),
-                server.session_id.clone(),
-                protocol::effective_protocol_version(
-                    &server.config,
-                    server
-                        .server_details
-                        .as_ref()
-                        .map(|details| details.protocol_version.as_str()),
-                ),
-            )
+        let Some(mut context) = self.tool_call_context(id) else {
+            return Err("Unknown MCP server".to_string());
         };
 
-        let client = self.ensure_http_client(id)?;
-
-        let mut request = apply_streamable_http_protocol_version_header(
-            apply_streamable_http_client_post_headers(client.post(base_url)),
-            Some(protocol_version.as_str()),
-        )
-        .body(payload);
-
-        if let Some(auth) = auth_header {
-            request = request.header("Authorization", auth);
+        transport_http::ensure_session_context(&mut context).await?;
+        self.update_tool_call_session(id, context.session_id.clone(), None);
+        if let Some(server) = self.server_mut(id) {
+            server.negotiated_protocol_version = context.negotiated_protocol_version.clone();
         }
-        if let Some(session_id) = session_id {
-            request = request.header("mcp-session-id", session_id);
-        }
-
-        let response = request.send().await.map_err(|err| err.to_string())?;
-        if !response.status().is_success() {
-            return Err(format!("HTTP error: {}", response.status()));
-        }
-
-        if let Some(session_id) = response
-            .headers()
-            .get("mcp-session-id")
-            .and_then(|value| value.to_str().ok())
-            .map(|value| value.to_string())
-        {
-            if let Some(server) = self.server_mut(id) {
-                server.session_id = Some(session_id);
-            }
-        }
-
         Ok(())
     }
 
@@ -1460,260 +856,24 @@ impl McpClientManager {
         id: &str,
         request: RequestFromClient,
     ) -> Result<ServerMessage, String> {
-        let (base_url, auth_header, session_id, request_id, protocol_version) = {
-            let Some(server) = self.server_mut(id) else {
-                return Err("Unknown MCP server".to_string());
-            };
-            let base_url = match require_http_base_url(&server.config) {
-                Ok(base_url) => base_url,
-                Err(err) => return Err(err),
-            };
-            let request_id = server.streamable_http_request_id;
-            server.streamable_http_request_id = server.streamable_http_request_id.saturating_add(1);
-            (
-                base_url,
-                server.auth_header.clone(),
-                server.session_id.clone(),
-                request_id,
-                protocol::effective_protocol_version(
-                    &server.config,
-                    server
-                        .server_details
-                        .as_ref()
-                        .map(|details| details.protocol_version.as_str()),
-                ),
-            )
+        self.ensure_http_client(id)?;
+
+        let Some(mut context) = self.tool_call_context(id) else {
+            return Err("Unknown MCP server".to_string());
         };
 
-        let message = ClientMessage::from_message(
-            MessageFromClient::RequestFromClient(request),
-            Some(RequestId::Integer(request_id as i64)),
+        let response = transport_http::send_request_with_context(
+            &mut context,
+            request,
+            self.server_request_tx.clone(),
         )
-        .map_err(|err| err.to_string())?;
-
-        let payload = serde_json::to_string(&message).map_err(|err| err.to_string())?;
-
-        let client = self.ensure_http_client(id)?;
-        debug!(
-            server_id = %id,
-            request_id,
-            url = %base_url,
-            "Sending MCP HTTP request"
-        );
-        let mut request = apply_streamable_http_protocol_version_header(
-            apply_streamable_http_client_post_headers(client.post(base_url)),
-            Some(protocol_version.as_str()),
-        )
-        .body(payload);
-
-        if let Some(auth) = auth_header {
-            request = request.header("Authorization", auth);
-        }
-        if let Some(session_id) = session_id {
-            request = request.header("mcp-session-id", session_id);
-        }
-
-        let response = request.send().await.map_err(|err| err.to_string())?;
-        debug!(
-            server_id = %id,
-            status = %response.status(),
-            "Received MCP HTTP response"
-        );
-        if !response.status().is_success() {
-            return Err(format!("HTTP error: {}", response.status()));
-        }
-
-        let session_id = response
-            .headers()
-            .get("mcp-session-id")
-            .and_then(|value| value.to_str().ok())
-            .map(|value| value.to_string());
-
-        let content_type = response
-            .headers()
-            .get(reqwest::header::CONTENT_TYPE)
-            .and_then(|value| value.to_str().ok())
-            .unwrap_or("")
-            .to_string();
-
-        let request_tx = self.server_request_tx.clone();
-        let server_message = if is_event_stream_content_type(&content_type) {
-            read_sse_response_messages(response, id, request_tx).await?
-        } else {
-            let body = response.bytes().await.map_err(|err| err.to_string())?;
-            serde_json::from_slice::<ServerMessage>(&body).map_err(|err| err.to_string())?
-        };
-
+        .await?;
+        self.update_tool_call_session(id, context.session_id.clone(), None);
         if let Some(server) = self.server_mut(id) {
-            if session_id.is_some() {
-                server.session_id = session_id;
-            }
+            server.negotiated_protocol_version = context.negotiated_protocol_version.clone();
         }
-
-        Ok(server_message)
+        Ok(response)
     }
-}
-
-fn sse_data_payload(line: &str) -> Option<&str> {
-    line.strip_prefix("data:").map(str::trim)
-}
-
-fn is_event_stream_content_type(content_type: &str) -> bool {
-    content_type
-        .split(';')
-        .next()
-        .map(str::trim)
-        .is_some_and(|value| value.eq_ignore_ascii_case("text/event-stream"))
-}
-
-async fn read_sse_response_messages(
-    response: reqwest::Response,
-    server_id: &str,
-    request_tx: Option<mpsc::UnboundedSender<McpServerRequest>>,
-) -> Result<ServerMessage, String> {
-    let mut stream = response.bytes_stream();
-    let mut buffer = SseLineBuffer::default();
-
-    while let Some(chunk) = stream.next().await {
-        let chunk = chunk.map_err(|err| err.to_string())?;
-        for line in buffer.push(&chunk) {
-            let Some(payload) = sse_data_payload(&line) else {
-                continue;
-            };
-            if payload.is_empty() {
-                continue;
-            }
-
-            let message =
-                serde_json::from_str::<ServerMessage>(payload).map_err(|err| err.to_string())?;
-            match message {
-                ServerMessage::Response(_) | ServerMessage::Error(_) => return Ok(message),
-                ServerMessage::Request(request) => {
-                    if let Some(tx) = request_tx.as_ref() {
-                        let _ = tx.send(McpServerRequest {
-                            server_id: server_id.to_string(),
-                            request,
-                        });
-                    }
-                }
-                ServerMessage::Notification(_) => {}
-            }
-        }
-    }
-
-    for line in buffer.finish() {
-        let Some(payload) = sse_data_payload(&line) else {
-            continue;
-        };
-        if payload.is_empty() {
-            continue;
-        }
-        let message =
-            serde_json::from_str::<ServerMessage>(payload).map_err(|err| err.to_string())?;
-        match message {
-            ServerMessage::Response(_) | ServerMessage::Error(_) => return Ok(message),
-            ServerMessage::Request(request) => {
-                if let Some(tx) = request_tx.as_ref() {
-                    let _ = tx.send(McpServerRequest {
-                        server_id: server_id.to_string(),
-                        request,
-                    });
-                }
-            }
-            ServerMessage::Notification(_) => {}
-        }
-    }
-
-    Err("Empty event-stream response.".to_string())
-}
-
-fn spawn_streamable_http_listener(
-    client: reqwest::Client,
-    server_id: String,
-    base_url: Option<String>,
-    auth_header: Option<String>,
-    session_id: Option<String>,
-    request_tx: mpsc::UnboundedSender<McpServerRequest>,
-    protocol_version: Option<String>,
-) {
-    let Some(base_url) = base_url else {
-        return;
-    };
-
-    tokio::spawn(async move {
-        let mut request = apply_streamable_http_protocol_version_header(
-            client.get(&base_url).header("Accept", "text/event-stream"),
-            protocol_version.as_deref(),
-        );
-
-        if let Some(auth) = auth_header {
-            request = request.header("Authorization", auth);
-        }
-        if let Some(session_id) = session_id {
-            request = request.header("mcp-session-id", session_id);
-        }
-
-        let response = match request.send().await {
-            Ok(response) => response,
-            Err(_) => return,
-        };
-
-        if !response.status().is_success() {
-            return;
-        }
-
-        let content_type = response
-            .headers()
-            .get(reqwest::header::CONTENT_TYPE)
-            .and_then(|value| value.to_str().ok())
-            .unwrap_or("");
-        if !is_event_stream_content_type(content_type) {
-            return;
-        }
-
-        let mut stream = response.bytes_stream();
-        let mut buffer = SseLineBuffer::default();
-
-        while let Some(chunk) = stream.next().await {
-            let chunk = match chunk {
-                Ok(chunk) => chunk,
-                Err(_) => return,
-            };
-            for line in buffer.push(&chunk) {
-                let Some(payload) = sse_data_payload(&line) else {
-                    continue;
-                };
-                if payload.is_empty() {
-                    continue;
-                }
-                if let Ok(ServerMessage::Request(request)) =
-                    serde_json::from_str::<ServerMessage>(payload)
-                {
-                    let _ = request_tx.send(McpServerRequest {
-                        server_id: server_id.clone(),
-                        request,
-                    });
-                }
-            }
-        }
-
-        for line in buffer.finish() {
-            let Some(payload) = sse_data_payload(&line) else {
-                continue;
-            };
-            if payload.is_empty() {
-                continue;
-            }
-            if let Ok(ServerMessage::Request(request)) =
-                serde_json::from_str::<ServerMessage>(payload)
-            {
-                let _ = request_tx.send(McpServerRequest {
-                    server_id: server_id.clone(),
-                    request,
-                });
-            }
-        }
-    });
 }
 
 #[derive(Clone)]
@@ -1751,21 +911,6 @@ pub struct McpServerRequestContext {
     http_client: Option<reqwest::Client>,
     client: Option<Arc<StdioClient>>,
     negotiated_protocol_version: Option<String>,
-}
-
-trait StreamableHttpContext {
-    fn config(&self) -> &McpServerConfig;
-    fn auth_header(&self) -> Option<&String>;
-    fn session_id(&self) -> Option<&String>;
-    fn http_client(&self) -> Option<&reqwest::Client>;
-    fn set_session_id(&mut self, session_id: Option<String>);
-    fn next_request_id(&mut self) -> i64;
-    fn negotiated_protocol_version(&self) -> Option<&str>;
-    fn set_negotiated_protocol_version(&mut self, protocol_version: Option<String>);
-
-    fn effective_protocol_version(&self) -> String {
-        protocol::effective_protocol_version(self.config(), self.negotiated_protocol_version())
-    }
 }
 
 impl StreamableHttpContext for McpToolCallContext {
@@ -1890,10 +1035,7 @@ impl McpClientManager {
             http_client: server.http_client.clone(),
             client: server.client.clone(),
             streamable_http_request_id: 0,
-            negotiated_protocol_version: server
-                .server_details
-                .as_ref()
-                .map(|details| details.protocol_version.clone()),
+            negotiated_protocol_version: server.negotiated_protocol_version.clone(),
         })
     }
 
@@ -1912,10 +1054,7 @@ impl McpClientManager {
             http_client: server.http_client.clone(),
             client: server.client.clone(),
             streamable_http_request_id: 0,
-            negotiated_protocol_version: server
-                .server_details
-                .as_ref()
-                .map(|details| details.protocol_version.clone()),
+            negotiated_protocol_version: server.negotiated_protocol_version.clone(),
         })
     }
 
@@ -1933,10 +1072,7 @@ impl McpClientManager {
             session_id: server.session_id.clone(),
             http_client: server.http_client.clone(),
             client: server.client.clone(),
-            negotiated_protocol_version: server
-                .server_details
-                .as_ref()
-                .map(|details| details.protocol_version.clone()),
+            negotiated_protocol_version: server.negotiated_protocol_version.clone(),
         })
     }
